tests: Fix warnings due to implicit int conversion

+++ orangecanvas/canvas/items/tests/test_annotationitem.py
@@ -58,7 +58,7 @@ class TestAnnotationItem(TestItems):
         def advance():
             clock = time.process_time() * 10
             item.setLineWidth(5 + math.sin(clock) * 5)
-            item.setColor(QColor(Qt.red).lighter(100 + 30 * math.cos(clock)))
+            item.setColor(QColor(Qt.red).lighter(100 + int(30 * math.cos(clock))))
 
         timer = QTimer(item, interval=10)
         timer.timeout.connect(advance)
